multi: Bump to spec v1.4.0
This bumps the rosetta-sdk-go to v0.3.0, the spec version to v1.4.0 and
the dcrros version to v0.2.0.

This is done in preparation for implementing the new Construction API
introduced in Rosetta v1.4.0.

The existing construction service implementation has been augmented with
the appropriate methods but they all currently return an
ErrUnimplemented error.