types: Add blockhash to reversed txs id
This modifies the transaction hash of reversed transactions to use
[prev_block_hash]:[tx_hash] format.

The purpose for this change is to ensure Rosetta blocks that encode a
disapproved Decred block cannot have duplicate transactions if the child
block contains a transaction also present in the parent block.

The Rosetta check:data validation asserts no two transactions in the
same block have the same hash. And it is valid according to current
consensus rules in Decred for a child block to include the same
transaction as its parent block if that child block disapproves the
parent.

Thefeore, after extensive discussions along with Rosetta's community
input, the approach chosen to represent such situations was to add the
block hash to tx identifiers that reverse the effects of a tx in the
regular tree of a parent disapproved block.