multi: Add DB tests
This adds tests to the current existing backenddb implementations. The
tests are done in a closed-box format, only using the interface-defined
functions.

This helps to ensure the correct behavior of the implementations
according to the current usage pattern of the functions within the
backend.

We also restrict implementations to only extending the tip with new
blocks and to not overwrite already processed blocks to ensure block
idempotency and prevent possible bugs that slip through to not put the
database in an invalid state.