multi: Check address via custom func
This changes the check done in /account/balance to use an exported
function from the types package (CheckRosettaAccount) instead of
dcrutil.DecodeAddress.

Some Rosetta accounts aren't represented by standard Decred addresses.
For example, pkscripts with a version > 0 or that are not standard
P2PKH/P2SH addresses.

These have a particular format that is not decodable by
dcrutil.DecodeAddress. Previously this could cause calls to
account/balance to fail for such addresses.