backend: Prevent processing blocks out of order
This adds a guard in the main processing function to avoid processing
blocks that don't extend the current tip, which would violate
assumptions.

It shoudn't be possible today to hit this condition outside of tests,
but this offers an additional guarantee in case there are undiscovered
bugs in the sync code.