backend: Avoid connecting earlier blocks
This modifies the logic in block notification processing such that we
don't start processing connected blocks until we receive a longer chain
than currently stored in the database.

This avoids unnecessary rollbacks in case we connect to a different node
that is behind in the chain and we start receiving block connected
notifications.