backend: Handle connecting to wrong dcrd node
This changes the server to handle connections to an unsuitable dcrd
node.

After verifying the node is unsuitable, operations that require access
to the node via the chain interface are disabled, ensuring the dcrros
server instance won't be put in an inconsitent state.

Several tests are added or modified to account for the new behavior.