backend: Switch to constructionTx on Construction API
This switches the unsigned and signed transactions of the Construction
API to use a custom structure called constructionTx.

Rosetta's Construction API requires most of its operations to be
performed in instances that are not actually connected to the larger
network. However, in order to acommplish some of its tasks, these
instances require additional transaction data, that must be specified
alongside the raw transaction being constructed.

In order to accomodate this additional data, a new structure named
'constructionTx' is introduced and all relevant Construction* calls are
modified to use this new structure instead of directly encoding the
transaction in wire format.

This new structure can be serialized and deserialized from an
hex-encoded, custom binary format.