backend: Allow using with offline nodes
This modifies the Server service to allow it connecting to a backing
dcrd node that does not have any blockchain data. Such nodes are usually
offline nodes, that are not connected to the larger network and are used
only for the offline Construction API operations.

This is implemented by waiting for peers for up to 10 seconds after the
server is initialized. If no peers are found, initialization proceeds
potentially skipping any blockchain sync or account preprocessing.