types: Use version 1 tx by default on conversions
This fixes a bug that created transactions with tx version 0 if the
metadata for tx version was unspecified. This could prevent transactions
generated by the Construction API from being mined and relayed.

We fix this by ensuring transactions generated by RosettaOpsToTx are
initialized by the standard wire.NewMsgTx() function.

Additionally, we switch some errors to variables for better
identification and add tests to assert edge cases of the RosettaOpsToTx
function.