backend: Add rpctest-based tests
This adds tests for the Data and Construction APIs using dcrd's rpctest
framework. This offers greater assurance that the overall integration
between dcrros and dcrd nodes are working as expected.

The Github Actions CI workflow has been updated to checkout and install
the latest stable dcrd version in order for the tests to run in that CI.