backend: Track account utxos
Rosetta spec v1.4.8 introduced the necessity to track per-account
unspent coins for the chain, in order to reply to the /account/coins
endpoint.

This commit adds per-account utxo tracking and implements the
/account/coins to correctly return such list.

Several cases need to be handled in order to properly track utxos:

- Standard creation (outputs) and spending (inputs) of utxos.
- Reversal of operations when a block disapproves its parent.
- Reversal of operations when reorganizing the chain.

A set of unit tests is included to exercise the expected behavior under
various circumstances. The balances rpctest is also modified to assert
the correct coins for the address under test are returned.