e2etest: Introduce package
This adds the first version of the full End-to-End test suite to dcrros.

This test suite aims to exercise a large portion of the features found
in the mainnet Decred chain by generating an extensive simnet, then
syncing a dcrros instance to it and then performing the rosetta-cli
check:data and check:construction tests.