Multi: Update to latest decred dependencies
This updates the decred module dependencies to their latest versions and
the dockerfiles to use the v1.7.0 release of the decred binaries.

The major change was the switch from dcrutil.Address to
stdaddress.Address to hold standard network addresses.

Another significant change is assuming the autorevocations agenda has
passed as required for running the E2E test in simnet.